pyproject.toml
[project]
name = "onnx-vehicle-plate-recognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "onnx>=1.18.0",
    "onnxruntime-gpu[cuda,cudnn]>=1.22.0",
    "opencv-contrib-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pyyaml>=6.0.2",
]
